Now its Bootstrap 3 compatable
Never mind that last fork, I didn't realize the modal didn't work either so I here are all the changes. The only thing I couldn't figure out with your code was adding the class "img-responsive" to the image after its uploaded. If not it will not fit well if to large. If you could add that. I could have done this with jQuery but I wasn't sure with your code.